feat: updated model, added id abstraction

+++ items-marketplace/src/main/java/me/darkovrbaski/items/marketplace/model/AccumulatedOrders.java
@@ -1,12 +1,13 @@
 package me.darkovrbaski.items.marketplace.model;
 
+import java.math.BigDecimal;
 import lombok.Value;
 
 @Value
 public class AccumulatedOrders {
 
-  double price;
+  BigDecimal price;
 
-  double quantity;
+  BigDecimal quantity;
 
 }
