+++ items-marketplace/src/main/java/me/darkovrbaski/items/marketplace/model/ArticleItem.java
@@ -6,6 +6,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,7 +21,7 @@ public class ArticleItem {
 
   @Positive
   @Column(nullable = false)
-  double quantity;
+  BigDecimal quantity;
 
   @ManyToOne(fetch = FetchType.EAGER)
   @JoinColumn(name = "article_id")
